Combine test_regions with test_platform
